Consolidate VCS solver source files
Code for the VCS_SOLVE class, which was split up among 12 source files,
is now consolidated into just vcs_solve.cpp and vcs_solve_TP.cpp.

This reduces the amount of time spent compiling the VCS solver by ~50%.
