+++ include/cantera/equil/vcs_species_thermo.h
@@ -6,7 +6,7 @@
 #ifndef VCS_SPECIES_THERMO_H
 #define VCS_SPECIES_THERMO_H
 
-#include <cstdlib>
+#include "cantera/equil/vcs_defs.h"
 
 namespace Cantera
 {
@@ -81,7 +81,23 @@ public:
     //! parameter that is used in the VCS_SSVOL_CONSTANT model.
     double SSStar_Vol0;
 
-    VCS_SPECIES_THERMO();
+    VCS_SPECIES_THERMO()
+         : IndexPhase(0)
+         , IndexSpeciesPhase(0)
+         , OwningPhase(0)
+         , SS0_Model(VCS_SS0_CONSTANT)
+         , SS0_feSave(0.0)
+         , SS0_TSave(-90.0)
+         , SS0_T0(273.15)
+         , SS0_H0(0.0)
+         , SS0_S0(0.0)
+         , SS0_Cp0(0.0)
+         , SS0_Pref(1.01325E5)
+         , SSStar_Model(VCS_SSSTAR_CONSTANT)
+         , SSStar_Vol_Model(VCS_SSVOL_IDEALGAS)
+         , SSStar_Vol0(-1.0)
+    {
+    }
 };
 
 }
